Tab and donut
I create a tab navigation and a donut chart. The donut chart has the
information hardcoded, but it should really pull the data from the CSV
files in the ‘about-team-files’ folder. Also, there are some images for
each team member in that same folder, which should be shown in the
center of the donut chart.